EnvmapImportanceSampling Extension Header+Impl
