+++ examples_tests/22.RaytracedAO/CMakeLists.txt
@@ -31,6 +31,7 @@ endif()
 
 set(EXTRA_SOURCES
 	../../src/nbl/ext/DebugDraw/CDraw3DLine.cpp
+	../../src/nbl/ext/EnvmapImportanceSampling/EnvmapImportanceSampling.cpp
 	Renderer.cpp
 	CommandLineHandler.cpp
 )
