+++ examples_tests/22.RaytracedAO/Renderer.h
@@ -7,6 +7,7 @@
 #undef PI
 
 #include "nbl/ext/MitsubaLoader/CMitsubaLoader.h"
+#include "nbl/ext/EnvmapImportanceSampling/EnvmapImportanceSampling.h"
 
 #include <ISceneManager.h>
 
@@ -246,48 +247,9 @@ class Renderer : public nbl::core::IReferenceCounted, public nbl::core::Interfac
 		nbl::core::smart_refctd_ptr<nbl::video::IGPURenderpassIndependentPipeline> blendEnvPipeline;
 		nbl::core::smart_refctd_ptr<nbl::video::IGPUDescriptorSetLayout> blendEnvDescriptorSet;
 		nbl::core::smart_refctd_ptr<nbl::video::IGPUMeshBuffer> blendEnvMeshBuffer;
-
-		// Shader and Resources for Generating Luminance MipMaps from EnvMap
-		struct ComputeWarpMapInfo
-		{
-			uint32_t MipCountEnvmap;
-			uint32_t MipCountLuminance;
-			uint32_t lumaMipMapGenWorkgroupDimension;
-			uint32_t warpMapGenWorkgroupDimension;
-		};
 		
-		static constexpr uint32_t MaxMipCountLuminance = 13u;
-		static constexpr uint32_t DefaultLumaMipMapGenWorkgroupDimension = 16u;
-		static constexpr uint32_t DefaultWarpMapGenWorkgroupDimension = 16u;
+		nbl::ext::EnvmapImportanceSampling::EnvmapImportanceSampling m_envMapImportanceSampling;
 
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUImageView> m_luminanceBaseImageView;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUImageView> m_luminanceMipMaps[MaxMipCountLuminance];
-		uint32_t m_lumaWorkGroups[2];
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUDescriptorSetLayout> m_lumaDSLayout;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUDescriptorSet> m_lumaDS[MaxMipCountLuminance - 1];
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUPipelineLayout> m_lumaPipelineLayout;
-		nbl::core::smart_refctd_ptr<IGPUSpecializedShader> m_lumaGPUShader;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUComputePipeline> m_lumaPipeline;
-
-		// Shader and Resources for EnvironmentalMap Sample Warping
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUImageView> m_warpMap; // Warps Sample based on EnvMap Luminance
-
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUDescriptorSetLayout> m_warpDSLayout;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUDescriptorSet> m_warpDS;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUPipelineLayout> m_warpPipelineLayout;
-		nbl::core::smart_refctd_ptr<IGPUSpecializedShader> m_warpGPUShader;
-		nbl::core::smart_refctd_ptr<nbl::video::IGPUComputePipeline> m_warpPipeline;
-		
-		ComputeWarpMapInfo initWarpingResources(
-			nbl::core::smart_refctd_ptr<nbl::video::IGPUImageView> envmap, 
-			uint32_t lumaMipMapGenWorkgroupDimension = DefaultLumaMipMapGenWorkgroupDimension,
-			uint32_t warpMapGenWorkgroupDimension = DefaultWarpMapGenWorkgroupDimension);
-		void deinitWarpingResources();
-
-		// returns if RIS should be enabled based on variance calculations
-		bool computeWarpMap(float envMapRegularizationFactor, ComputeWarpMapInfo info);
-
-		ComputeWarpMapInfo m_computeWarpMapInfo;
 		std::future<bool> compileShadersFuture;
 };
